backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
httpx>=0.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
